Cleanup and restore int96timestamps for this PR.

+++ cpp/include/cudf/types.hpp
@@ -216,7 +216,7 @@ enum class type_id : int32_t {
   TIMESTAMP_MILLISECONDS,  ///< point in time in milliseconds since Unix Epoch in int64
   TIMESTAMP_MICROSECONDS,  ///< point in time in microseconds since Unix Epoch in int64
   TIMESTAMP_NANOSECONDS,   ///< point in time in nanoseconds since Unix Epoch in int64
-  DURATION_DAYS,           ///< time interval of days in int64
+  DURATION_DAYS,           ///< time interval of days in int32
   DURATION_SECONDS,        ///< time interval of seconds in int64
   DURATION_MILLISECONDS,   ///< time interval of milliseconds in int64
   DURATION_MICROSECONDS,   ///< time interval of microseconds in int64
